requirements.txt
asgiref==3.5.2
certifi==2020.4.5.1
channels==2.1.7
channels-redis==2.4.2
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0
django-allauth==0.51.0
django-braces==1.14.0
django-channels==0.7.0
django-ckeditor==6.4.2
django-cors-headers==3.5.0
django-heroku==0.3.1
django-memcache-status==2.2
django-rest-knox==4.2.0
django-restframework==0.0.1
django-widget-tweaks==1.4.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
domain-coordinator==0.9.2
drf-yasg==1.21.4
facebook-sdk==3.1.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.1
pipenv==2020.11.15
platformdirs==2.5.1
protobuf==3.13.0
psycopg2==2.9.3
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
SQLAlchemy==1.4.40
sqlparse==0.3.1
twilio==7.14.0
validate-email==1.3
whitenoise==5.2.0
